Validate input in inverse captcha module (day 1)
The 'solveCaptcha' method does now check if the string received as an
input is a valid sequence or not. If the string contains any non-numeric
character, the function will print an error message telling the user the
input should contain numbers only.

New tests have been also added to test this new feature.