Added node script to run functions in modules
I've created a script that receives three values as a parameter via
Node's CLI:

- Day (a number between 1 and 25)
- Part (it can be either 1 or 2)
- Input (the input which will be received by the function)

The two first parameters will be used to determine which of the modules
will be run. The path of the module will be different depending of the
values you pass here.

The input passed as a third argument will be passed to the function
defined in the specified module, and the script will display the result
returned by that function for the input you chose.